Create Script.js
In this example, the getUserMedia function is used to request access to the user's camera and microphone, and the resulting stream is played in a video element. A MediaRecorder instance is then created to record the video, and the recorded data is stored in an array of Blob objects. When the recording is stopped, the recorded chunks are combined into a single Blob representing the complete video file. Finally, a download link is created for the video file and added to the page, allowing the user to save the recorded video.